putSPCsに負の数を渡さないように修正
カーソルが画面外にあるとき負の個数のスペースを置こうとしてしまっていたのを直した
配列のサイズとして使う数を受け取る時などはsize_tを使うように直した
ついでに不等号が逆向きだったのを直しておいた

+++ line.h
@@ -42,11 +42,11 @@ void freeLine(Line *);
 void insertU32s(Line *, int, const InsertLine *, int);
 void deleteChars(Line *, int, int);
 int eraseInLine(Line *, int, int);
-int putU32s(Line *, int, const char32_t *, int, int, int, int);
-void putSPCs(Line *, int, int, int);
+int putU32s(Line *, int, const char32_t *, int, int, int, size_t);
+void putSPCs(Line *, int, int, size_t);
 CharCnt getCharCnt(const Line *, int);
 int findNextSGR(const Line *, int);
 
 const char *u8sToU32s(char32_t *,const char *, size_t);
-int u32slen(const char32_t *);
+size_t u32slen(const char32_t *);
 int u32swidth(const char32_t *, int);
